Split out menus in UI
