+++ src/ui/mod.rs
@@ -1,4 +1,9 @@
 mod uiapp;
 mod defines;
+mod ui_settings;
+mod ui_import_games;
+
 pub use uiapp::*;
 pub use defines::*;
+pub use ui_settings::*;
+pub use ui_import_games::*;
